[기능 구현] 퀘스트 진행, 퀘스트 완료 임시 구현, 메뉴에 수락 가능 목록 보기 추가

+++ app/src/main/java/com/example/hyojung/quest/Activity/LoginActivity.java
@@ -51,7 +51,7 @@ public class LoginActivity extends AppCompatActivity {
             public void onClick(View v) {
                 LoginQuery clientLoginQuery = new LoginQuery(0, "관리자", "관리자");
                 Intent intent = new Intent(instance, MainActivity.class);
-                intent.putExtra("kakaoID", 0);
+                intent.putExtra("kakaoID", (long)0);
                 intent.putExtra("kakaoNickName", "관리자");
                 intent.putExtra("kakaoProfileImage", "관리자");
                 JSONSendTask jsonTask = new JSONSendTask(clientLoginQuery);
